lb1 done

+++ src/modules/Ship.hpp
@@ -16,10 +16,10 @@ class Ship {
     Segment* segments;
 
    public:
-    Ship(size_t length);
+    Ship(int length);
     ~Ship();
 
-    void takeDamage(size_t segment_index, int damage);     // returns current hp of segment
+    void takeDamage(size_t segment_index, int damage);     // deals damage to segment by index
     int getSegmentHP(size_t index) const;                  // returns hp of segment by index
     ShipSegmentState getSegmentState(size_t index) const;  // returns state of segment by index
     size_t getLength() const;                              // returns length of ship
